requirements.txt
streamlit==1.31.0
pandas==2.1.0
PyPDF2==3.0.1
python-pptx==0.6.21
python-dotenv==1.0.0
requests==2.31.0